requirements.txt
# Requirements for Advanced PDF Reader
PyMuPDF
Pillow
pywin32
pystray
requests
# tkinter is included with standard Python on Windows 
